run forge fmt

+++ src/tests/mock/MockWETH.sol
@@ -1,18 +1,18 @@
 // SPDX-License-Identifier: BUSL-1.1
 pragma solidity 0.8.19;
 
-import 'openzeppelin/contracts/token/ERC20/ERC20.sol';
+import "openzeppelin/contracts/token/ERC20/ERC20.sol";
 
 contract MockWETH is ERC20 {
-  constructor() ERC20('WETH', 'WETH') {}
+    constructor() ERC20("WETH", "WETH") {}
 
-  function deposit() external payable {
-    _mint(msg.sender, msg.value);
-  }
+    function deposit() external payable {
+        _mint(msg.sender, msg.value);
+    }
 
-  function withdraw(uint amount) external {
-    _burn(msg.sender, amount);
-    (bool success, ) = msg.sender.call{value: amount}(new bytes(0));
-    require(success, '!success');
-  }
+    function withdraw(uint256 amount) external {
+        _burn(msg.sender, amount);
+        (bool success,) = msg.sender.call{value: amount}(new bytes(0));
+        require(success, "!success");
+    }
 }
